Implemented AV2-182: Implement normalization feature -refactoring
